requirements.txt
tensorflow-gpu==2.9.1
keras==2.9.0
numpy==1.21.6
h5py==2.10.0
protobuf==3.19.6
tensorflow-io-gcs-filesystem==0.26.0
opencv-python==4.6.0.66
albumentations==1.3.1
scikit-image==0.19.3
pillow==9.5.0
imageio==2.5.0
imgaug==0.2.9
pandas==1.3.5
scikit-learn==1.0.2
matplotlib==3.5.3
tqdm==4.64.0
psutil==5.9.8
keras-flops==0.1.2
keras-segmentation==0.3.0
tensorboard==2.9.1